Otimizando arquivos

+++ Banco/Caelum.Banco.Interfaces/ITributavel.cs
@@ -2,6 +2,6 @@
 {
     public interface ITributavel
     {
-        double CalculaTributo();
+        decimal CalculaTributo();
     }
 }
