+++ Banco/Caelum.Banco.Services/TributosServices.cs
@@ -4,7 +4,7 @@ namespace Caelum.Banco.Services
 {
     public class TributosServices
     {
-        public double Total { get; private set; }
+        public decimal Total { get; private set; }
 
         public void Acumula(ITributavel it)
         {
